% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouper.R
\name{grouper}
\alias{grouper}
\title{Group together low frequency values of a categorical variable}
\usage{
grouper(data, categvar, metric, p = 0.03, suggestp = F, noutcat = NULL)
}
\arguments{
\item{data}{a dataframe}

\item{categvar}{a string containing the name of the categorical variable in d that will be grouped}

\item{metric}{a string containing the name of a numeric variable in d which will be used to determine which values of x to map into <other>}

\item{p}{a threshold indicating the percent of metric under which values of x will be mapped to <other>}
}
\value{
a data.frame with one new column, x.new, where x is the name of the column to be grouped.
}
\description{
The intenttion is to take a categorical variable with many values and transform it into a another categorical variable with a manageable number of values. For example, convert zipcode with 32000 values into a variable with 250 values. THe retained values will map 1-1 with the original values, except for <other>, which will map to all the low frequency original values.
}
\examples{
y = grouper(data = mydata, categvar='Species', metric='population', p=.03)
}
\author{
Rahul Dodhia
}

