% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{HelperFunctions-package}
\alias{HelperFunctions}
\alias{HelperFunctions-package}
\title{Helper Functions and Data (Avvo edition)}
\description{
Some functions to make common tasks more efficient. Also includes some datasets for easy lookup.
}
\details{
The code is stored on Gitlab, and freely available to anyone with access to the project to add to it.
To update or install, use \code{install_git('git@gitlab.corp.avvo.com:analytics/R_Functions.git')}
To get more information about the package, including a list of functions and datasets, use \code{library(help="HelperFunctions")}
}
\author{
Rahul Dodhia
}

