% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invlogit.R
\name{invlogit}
\alias{invlogit}
\title{Inverse logit.}
\usage{
invlogit(x)
}
\arguments{
\item{x}{a numeric vector or scalar}
}
\value{
Vector or scalar of the inverse logit of x
}
\description{
Inverse logit.
}
\examples{
y = invlogit(x = c(-4,0,100,3))
}
\author{
Rahul Dodhia
}

