% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.R
\name{km}
\alias{km}
\title{Wrapper for k-means.}
\usage{
km(x, minnum = 4, ...)
}
\arguments{
\item{x}{numeric matrix of data, or an object that can be coerced to such a matrix (such as a numeric vector or a data frame with all numeric columns).}

\item{...}{any arguments that can be passed to kmeans().}
}
\value{
List with 2 elements.
\describe{
 \item{out}{contains standard kmeans() output,}
 \item{realcenters}{gives the centers on the original, non-standardized variables.}
}
}
\description{
Makes all columns numeric, removes row with missing values, standardizes columns to mean 1 and standard deviation 0.
}
\examples{
y = km(m, centers=10, iter.max=20)
}
\author{
Rahul Dodhia
}

