% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onelineq.R
\name{onelineq}
\alias{onelineq}
\title{Helper function for readimpala()}
\usage{
onelineq(sqlfile, type = "file")
}
\arguments{
\item{sqlfile}{file containing sql query}

\item{x}{character vector or list of character vectors}
}
\value{
character variable for use in readimpala()
}
\description{
Cleans up sql query for use in readimpala(). Not usually run by itself, called by readimpala().
}
\examples{
x = readimpala('~/project_folder/example.sql','rdodhia')
}
\author{
Rahul Dodhia
}

