% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppc}
\alias{ppc}
\title{PPC multipliers for Avvo markets}
\format{A data.table with 470098 rows and 2 variables}
\source{
dm.lz_ad_pricing_ppc. (Switch to dm.ad_pricing_ppc)
}
\usage{
ppc
}
\description{

}
\details{
A dataset containing ppc values for each market
}
\keyword{datasets}

