% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{Read files}
\usage{
read(x)
}
\arguments{
\item{x}{file name, include file path}
}
\value{
data.table
}
\description{
Read files
}
\examples{
y = read('file.csv')
y = read('file.tab')
}

