% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readimpala.R
\name{readimpala}
\alias{readimpala}
\title{Reads data from impala, using your ODBC connection.}
\usage{
readimpala(file = "", query = "", user = "rdodhia",
  dsn = "Cloudera Impala", type = "file")
}
\arguments{
\item{file}{sql file}

\item{query}{a string with a sql query.}

\item{user}{kerberos username}

\item{dsn}{The name ofthe ODBC connection on your computer. Defaults to "Cloudera Impala".}
}
\value{
Data.table containing output of sql query
}
\description{
Reads data from impala, using your ODBC connection.
}
\details{
If a file and query are both given, the file will be used.
}
\examples{
x = readimpala('~/project_folder/example.sql','rdodhia')
}
\author{
Rahul Dodhia
}

