% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Standard error of the mean}
\usage{
se(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
A scalar
}
\description{
R has a built-in sd() function which calculates the standard deviation of a sample. This function is used to derive the s.e. of the mean. Missing values will be ignored.
}
\examples{
y = se(c(3,8,NA,3,1,10))
}
\author{
Rahul Dodhia
}

