% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todate.R
\name{todate}
\alias{todate}
\title{Add date formats to data}
\usage{
todate(d)
}
\arguments{
\item{d}{a data.table or data.frame}
}
\value{
the input data d with several columns appended
}
\description{
Appends several date formats to a data.table. Assumes that one of the columns in the input data is labeled "date" and is in Y-m-d format.
}
\author{
Rahul M Dodhia
}

