% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimws.R
\name{trimws}
\alias{trimws}
\title{Remove whitespace}
\usage{
trimws(x)
}
\arguments{
\item{x}{A string}
}
\value{
String whithout leading and trailingwhitespace
}
\description{
Removes whitespace before and after a string
}
\examples{
y = trim('  hello  ')
sapply(x,trim) #x is a vector of strings
}
\author{
Rahul Dodhia
}

